On Verify: CreateMrStock (See CreateDeliverySKUS) You need to Create stock with the same sku numbers as the Sales Order Items. This stock needs to be created in the receiving bay and then it needs to be putaway by the user (They do this with the PDTs) (Do not track these putaways for now) Once the Sales Return has been verified: repo.update_with_document_number('doc_seqs_sales_return_number', id)
Once the Sales Return has been verified allow the user to print the SKU Barcode label from the action menu. See Delivery Line items
After the Sales Return has been verified allow the user to update the price 'inline' (in the grid), but do not allow the user to further edit the sales return.
Sales Return can not be reopened because it has created stock.

Once the Sales Return has been verified the user with the 'integrate' permission can complete sales return.

 Please ignore this:
After the Sales Return has been verified allow the user to update the price 'inline' (in the grid), but do not allow the user to further edit the sales return.

The price of the item needs to be the exact same price as set on the sales order.

Important:

    Only Allow shipped Sales Orders for creating new Sales Returns.
    Show the Sales Order Number in the Sales return preselect dropdown and not the ID
    On create new sales return item the grid is not populated with the product variant code, please pull through the value.